move validation tests into separate module
