+++ tests/unit/test_config.py
@@ -1,19 +1,10 @@
-import errno
 from pathlib import Path
 
 import pytest
 
-from linker.configuration import (
-    DEFAULT_ENVIRONMENT,
-    ENVIRONMENT_ERRORS_KEY,
-    INPUT_DATA_ERRORS_KEY,
-    PIPELINE_ERRORS_KEY,
-    Config,
-)
+from linker.configuration import DEFAULT_ENVIRONMENT, Config
 from linker.step import Step
-from linker.utilities import paths
 from linker.utilities.data_utils import load_yaml
-from tests.unit.conftest import check_expected_validation_exit
 
 
 @pytest.mark.parametrize("requires_spark", [True, False])
@@ -165,271 +156,3 @@ def test_get_implementation_specific_configuration(
         config.pipeline["steps"]["step_2"]["implementation"]["configuration"] = step_2_config
     assert config.get_implementation_specific_configuration("step_1") == step_1_config
     assert config.get_implementation_specific_configuration("step_2") == step_2_config
-
-
-####################
-# validation tests #
-####################
-
-
-@pytest.mark.skip(reason="TODO [MIC-4735]")
-def test_batch_validation():
-    pass
-
-
-# Pipeline validations
-@pytest.mark.parametrize(
-    "pipeline, expected_msg",
-    [
-        # missing 'implementation' key
-        (
-            "missing_implementation_pipeline.yaml",
-            {
-                PIPELINE_ERRORS_KEY: {
-                    "step step_1": ["Does not contain an 'implementation'."]
-                },
-            },
-        ),
-        # missing implementation 'name' key
-        (
-            "missing_implementation_name_pipeline.yaml",
-            {
-                PIPELINE_ERRORS_KEY: {
-                    "step step_1": ["The implementation does not contain a 'name'."]
-                },
-            },
-        ),
-        # steps are out of order
-        (
-            "out_of_order_pipeline.yaml",
-            {
-                PIPELINE_ERRORS_KEY: {
-                    "development": [
-                        "- Step 1: the pipeline schema expects step step_1 but "
-                        "the provided pipeline specifies step_2. Check step order "
-                        "and spelling in the pipeline configuration yaml. "
-                        "- Step 2: the pipeline schema expects step step_2 but "
-                        "the provided pipeline specifies step_1. Check step order "
-                        "and spelling in the pipeline configuration yaml."
-                    ],
-                    "pvs_like_case_study": [
-                        "- Expected 1 steps but found 2 implementations. Check "
-                        "that all steps are accounted for \\(and there are no extraneous "
-                        "ones\\) in the pipeline configuration yaml."
-                    ],
-                },
-            },
-        ),
-        # missing a step
-        (
-            "missing_step_pipeline.yaml",
-            {
-                PIPELINE_ERRORS_KEY: {
-                    "development": [
-                        "- Expected 2 steps but found 1 implementations. Check that "
-                        "all steps are accounted for \\(and there are no extraneous "
-                        "ones\\) in the pipeline configuration yaml.",
-                    ],
-                    "pvs_like_case_study": [
-                        "- 'Step 1: the pipeline schema expects step 'pvs_like_case_study' "
-                        "but the provided pipeline specifies 'step_2'. Check step order "
-                        "and spelling in the pipeline configuration yaml.'",
-                    ],
-                },
-            },
-        ),
-    ],
-)
-def test_pipeline_validation(pipeline, expected_msg, test_dir, caplog, mocker):
-    mocker.patch(
-        "linker.configuration.Config._determine_if_spark_is_required", return_value=False
-    )
-    config_params = {
-        "pipeline_specification": Path(f"{test_dir}/{pipeline}"),
-        "input_data": Path(f"{test_dir}/input_data.yaml"),
-        "computing_environment": Path(f"{test_dir}/environment.yaml"),
-    }
-
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg=expected_msg,
-    )
-
-
-def test_unsupported_step(test_dir, caplog, mocker):
-    mocker.patch("linker.implementation.Implementation._load_metadata")
-    mocker.patch("linker.implementation.Implementation._get_container_full_stem")
-    mocker.patch("linker.implementation.Implementation.validate", return_value=[])
-    config_params = {
-        "pipeline_specification": Path(f"{test_dir}/bad_step_pipeline.yaml"),
-        "input_data": Path(f"{test_dir}/input_data.yaml"),
-        "computing_environment": Path(f"{test_dir}/environment.yaml"),
-    }
-
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            PIPELINE_ERRORS_KEY: {
-                "development": [
-                    "- Expected 2 steps but found 1 implementations. Check that "
-                    "all steps are accounted for \\(and there are no extraneous "
-                    "ones\\) in the pipeline configuration yaml.",
-                ],
-                "pvs_like_case_study": [
-                    "- 'Step 1: the pipeline schema expects step 'pvs_like_case_study' "
-                    "but the provided pipeline specifies 'foo'. Check step order "
-                    "and spelling in the pipeline configuration yaml.'",
-                ],
-            }
-        },
-    )
-
-
-def test_unsupported_implementation(test_dir, caplog, mocker):
-    mocker.patch("linker.implementation.Implementation._load_metadata")
-    mocker.patch("linker.implementation.Implementation._get_container_full_stem")
-    mocker.patch("linker.implementation.Implementation.validate", return_value=[])
-    mocker.patch(
-        "linker.configuration.Config._determine_if_spark_is_required", return_value=False
-    )
-    config_params = {
-        "pipeline_specification": Path(f"{test_dir}/bad_implementation_pipeline.yaml"),
-        "input_data": Path(f"{test_dir}/input_data.yaml"),
-        "computing_environment": Path(f"{test_dir}/environment.yaml"),
-    }
-
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-
-    implementation_metadata = load_yaml(paths.IMPLEMENTATION_METADATA)
-    supported_implementations = (
-        str(list(implementation_metadata.keys())).replace("[", "\\[").replace("]", "\\]")
-    )
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            PIPELINE_ERRORS_KEY: {
-                "step step_1": [
-                    f"Implementation 'foo' is not supported. Supported implementations are: {supported_implementations}."
-                ],
-            }
-        },
-    )
-
-
-# Input data validations
-def test_bad_input_data(test_dir, caplog):
-    config_params = {
-        "pipeline_specification": f"{test_dir}/pipeline.yaml",
-        "input_data": f"{test_dir}/bad_columns_input_data.yaml",
-        "computing_environment": None,
-    }
-
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            INPUT_DATA_ERRORS_KEY: {
-                ".*/broken_file1.csv": [
-                    "- Data file .* is missing required column\\(s\\) .*"
-                ],
-                ".*/broken_file2.csv": [
-                    "- Data file .* is missing required column\\(s\\) .*"
-                ],
-            }
-        },
-    )
-
-
-def test_missing_input_file(test_dir, caplog):
-    config_params = {
-        "pipeline_specification": f"{test_dir}/pipeline.yaml",
-        "input_data": f"{test_dir}/missing_input_data.yaml",
-        "computing_environment": None,
-    }
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            INPUT_DATA_ERRORS_KEY: {
-                ".*/missing_file1.csv": ["File not found."],
-                ".*/missing_file2.csv": ["File not found."],
-            },
-        },
-    )
-
-
-# Environment validations
-def test_unsupported_container_engine(test_dir, caplog, mocker):
-    config_params = {
-        "pipeline_specification": f"{test_dir}/pipeline.yaml",
-        "input_data": f"{test_dir}/input_data.yaml",
-        "computing_environment": None,
-    }
-    mocker.patch(
-        "linker.configuration.Config._get_required_attribute",
-        side_effect=lambda _env, attribute: "foo"
-        if attribute == "container_engine"
-        else None,
-    )
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            "ENVIRONMENT ERRORS": {
-                "container_engine": ["The value 'foo' is not supported."],
-            },
-        },
-    )
-
-
-def test_missing_slurm_details(test_dir, caplog, mocker):
-    mocker.patch(
-        "linker.configuration.Config._get_required_attribute",
-        side_effect=lambda _env, attribute: "slurm"
-        if attribute == "computing_environment"
-        else "undefined",
-    )
-    config_params = {
-        "pipeline_specification": f"{test_dir}/pipeline.yaml",
-        "input_data": f"{test_dir}/input_data.yaml",
-        "computing_environment": None,
-    }
-    with pytest.raises(SystemExit) as e:
-        Config(**config_params)
-    check_expected_validation_exit(
-        error=e,
-        caplog=caplog,
-        error_no=errno.EINVAL,
-        expected_msg={
-            ENVIRONMENT_ERRORS_KEY: {
-                "slurm": [
-                    "The environment configuration file must include a 'slurm' key "
-                    "defining slurm resources if the computing_environment is 'slurm'."
-                ],
-            },
-        },
-    )
